app/requirements.txt
aiocontextvars==0.2.2
amqp==2.5.2
arrow==0.15.7
asgiref==3.2.5
billiard==3.6.3.0
blessed==1.17.8
certifi==2019.11.28
chardet==3.0.4
contextvars==2.4
Django==3.0.7
django-crispy-forms==1.9.0
django-extra-views==0.13.0
django-filter==2.2.0
django-picklefield==3.0.1
django-q==1.2.4
django-tables2==2.2.1
django-timezone-field==4.0
gunicorn==20.0.4
idna==2.9
immutables==0.11
importlib-metadata==1.5.0
kombu==4.6.8
marshmallow==3.5.1
pendulum==2.0.3
psycopg2==2.8.4
python-crontab==2.4.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pytzdata==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.2.4
zipp==3.1.0
